Fix delete & move icon
